--- 2. Data Center Cooling/README.md
@@ -1,20 +1,0 @@
-# Data Center Cooling
-![](https://s-media-cache-ak0.pinimg.com/originals/36/d1/87/36d18741bdd4d2ac0033c53bcc669148.jpg)
-
-Inspired by [DeepMind's realization](https://deepmind.com/blog/deepmind-ai-reduces-google-data-centre-cooling-bill-40/)
-
-This repository hold the development of a business question that can be solved with Reinforcement Learning : cooling data centers
-- The environment modelled with the fashion of OpenAI Gym's environments
-- Solving the problem with different RL algorithms (Q-Learning, Deep-Q-Learning, Policy Gradients)
-- A interactive Dash app to test the environment and the agents
-
-
-***
-## Data Center Cooling environment
-
-To try out the app, launch with ``python app.py`` and go to port ``localhost:8050``
-
-
-
-
-
